pyproject.toml
[tool.poetry]
name = "llm-server"
version = "0.0.1"
description = "A server for running LLM behind an API"
authors = ["MrFastwind <nivek1999@yahoo.it>"]
license = "MIT"
packages = [{include = "llmserver"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
transformers = "^4.29.2"
pydantic = "^1.10.8"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
uvicorn = "^0.22.0"
pyyaml = "^6.0"
faiss-cpu = "^1.7.4"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "llmserver.__main__:main"